+++ src/main/java/com/lamarrulla/proyectointegrador/repository/ModelRespository.java
@@ -3,5 +3,5 @@ package com.lamarrulla.proyectointegrador.repository;
 import com.lamarrulla.proyectointegrador.entity.Modelo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModelRespository extends JpaRepository<Modelo, Long> {
+public interface ModelRespository extends JpaRepository<Modelo, Integer> {
 }
